requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# LLM routing
litellm>=1.40.0
openai>=1.30.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# HTTP client with retry
httpx>=0.27.0
tenacity>=8.3.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Database
motor>=3.4.0          # async MongoDB
redis[hiredis]>=5.0.0 # async Redis

# Logging
loguru>=0.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0

# Optional: monitoring
# prometheus-fastapi-instrumentator>=7.0.0

## 
redis
python-dotenv